requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
sphinx==1.8.5
pokedex.py==1.1.2